backend/requirements.txt
flask
flask-cors
psycopg2
requests
python-dotenv